% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa-package.R
\docType{package}
\name{dyntaxa-package}
\alias{dyntaxa-package}
\alias{dyntaxa}
\title{Access to Dyntaxa data - the Swedish Taxonomic Database - from R}
\description{
The goal of the `dyntaxa` R package is to provide a read-only R package to interface with Dyntaxa - the taxonomic database of organisms in Sweden.
}
\details{
The Dyntaxa data contains information about 61,500 species occurring in Sweden. This 
includes about 95% of known multicellular species – remaining gaps mainly 
found among the fungi. 
The scope of organisms include multicellular species documented in Sweden 
and such unicellular species that are included in environmental monitoring 
by the Swedish EPA.
In addition to these species there are many names at other taxonomic levels, scientific synonyms, Swedish vernaculars.

Citation: Liljeblad J (2019). Dyntaxa. Svensk taxonomisk databas. ArtDatabanken. Checklist dataset https://doi.org/10.15468/j43wfc.
}
\author{
Markus Skyttner \email{markus.skyttner@nrm.se}
}
\keyword{package}
