% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_ng.R
\name{dyntaxa_archive}
\alias{dyntaxa_archive}
\title{Listing of available dyntaxa database archive files}
\usage{
dyntaxa_archive()
}
\description{
Listing of available dyntaxa database archive files
}
\details{
Source: https://archive.infrabas.se/dyntaxa/
}
