% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_ng.R
\name{dyntaxa_archive_diffs}
\alias{dyntaxa_archive_diffs}
\title{Deduplicates dyntaxa files downloadable from the archive mirror}
\usage{
dyntaxa_archive_diffs()
}
\value{
a tibble with an attribute "prune_cmd"
}
\description{
This function makes a HEAD request and based on content length,
it computes a diff.
}
\details{
a suggested command that can be used to prune dupes is
provided as an attribute attached to the result
}
