% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_children}
\alias{dyntaxa_children}
\title{Retrieve immediate descendents, ie only children at the rank one step below that of a given taxon}
\usage{
dyntaxa_children(taxon_id)
}
\arguments{
\item{taxon_id}{character string with taxon identifier/key}
}
\value{
tibble with the columns: childtaxa_id,  childtaxa_name,
}
\description{
Retrieve immediate descendents, ie only children at the rank one step below that of a given taxon
}
\examples{
\dontrun{
dyntaxa_children(dyntaxa_id_from_name("Agaricales"))
}
}
