% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_classification}
\alias{dyntaxa_classification}
\title{Retrieve the taxonomic "upstream" classification given a taxon identifier}
\usage{
dyntaxa_classification(taxon_id)
}
\arguments{
\item{taxon_id}{taxon identifier/key}
}
\value{
tibble with the columns: name, rank, and id.
}
\description{
Retrieve the taxonomic "upstream" classification given a taxon identifier
}
\examples{
\dontrun{
dyntaxa_classification(dyntaxa_id_from_name("Cervidae"))
}
}
