% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_ng.R
\name{dyntaxa_con}
\alias{dyntaxa_con}
\title{A connection to a full text searchable index of Dyntaxa}
\usage{
dyntaxa_con(db, refresh_url)
}
\arguments{
\item{db}{optional path to local database file (see dyntaxa_fts() for location)}

\item{refresh_url}{optional url for downloading data from non-default locations}
}
\description{
If a local database does not exist, this function downloads and creates it.
}
