% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_downstream}
\alias{dyntaxa_downstream}
\title{Retrieve all taxa "downstream" of a given taxon identifier, descending recursively through all available taxonomic ranks}
\usage{
dyntaxa_downstream(taxon_id)
}
\arguments{
\item{taxon_id}{taxon identifier/key}
}
\value{
tibble with the columns: childtaxa_id, childtaxa_name and rank.
}
\description{
Retrieve all taxa "downstream" of a given taxon identifier, descending recursively through all available taxonomic ranks
}
\examples{
\dontrun{
dyntaxa_downstream(dyntaxa_id_from_name("Asteraceae"))
}
}
