% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_dwca}
\alias{dyntaxa_dwca}
\title{Access internal package R object representing the darwin core archive data object for the Dyntaxa database}
\usage{
dyntaxa_dwca()
}
\value{
object with slots for the darwin core archive data
}
\description{
Access internal package R object representing the darwin core archive data object for the Dyntaxa database
}
\examples{
\dontrun{
d <- dyntaxa_dwca()
dt <- d$taxon_core
dt
}
}
