% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwca_export.R
\name{dyntaxa_export}
\alias{dyntaxa_export}
\title{Exports Dyntaxa data as a darwin core archive file}
\usage{
dyntaxa_export(outfile)
}
\arguments{
\item{outfile}{filename to use for the darwincore archive file export}
}
\value{
path to exported darwin core archive file
}
\description{
Exports Dyntaxa data as a darwin core archive file
}
\examples{
\dontrun{
 dyntaxa_export()
}
}
