% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_ng.R
\name{dyntaxa_fts}
\alias{dyntaxa_fts}
\title{Location on disk for dyntaxa SQLite3 database}
\usage{
dyntaxa_fts()
}
\value{
the path where the downloaded SQLite3 database will reside
}
\description{
Returns the location for the SQLite3 database
}
