% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_id_from_name}
\alias{dyntaxa_id_from_name}
\title{Get taxon identifiers from taxon name}
\usage{
dyntaxa_id_from_name(taxon_name)
}
\arguments{
\item{taxon_name}{Vector of taxon names}
}
\value{
vector of identifiers
}
\description{
Get taxon identifiers from taxon name
}
\examples{
\dontrun{
dyntaxa_id_from_name("Vulpes vulpes"))
}
}
