% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_init}
\alias{dyntaxa_init}
\title{Download Dyntaxa darwin core archive and 
prepare data for local use with this R package}
\usage{
dyntaxa_init(force = FALSE)
}
\arguments{
\item{force}{boolean to indicate a forced update}
}
\description{
Download Dyntaxa darwin core archive and 
prepare data for local use with this R package
}
