% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_ng.R
\name{dyntaxa_init_fts}
\alias{dyntaxa_init_fts}
\title{Download Dyntaxa darwin core archive and 
prepare data for local use with this R package}
\usage{
dyntaxa_init_fts()
}
\description{
Download Dyntaxa darwin core archive and 
prepare data for local use with this R package
}
