% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_links}
\alias{dyntaxa_links}
\title{Get links given a specific taxon identifier}
\usage{
dyntaxa_links(taxon_id)
}
\arguments{
\item{taxon_id}{taxon identifier/key}
}
\value{
tibble with info links
}
\description{
Get links given a specific taxon identifier
}
\examples{
\dontrun{
link <- dyntaxa_links("5000001")$links[1]
browseURL(link)
}
}
