% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_name_from_id}
\alias{dyntaxa_name_from_id}
\title{Get scientific name given key/identifier}
\usage{
dyntaxa_name_from_id(taxon_id)
}
\arguments{
\item{taxon_id}{search term for key/identifier}
}
\value{
vector of names
}
\description{
Get scientific name given key/identifier
}
\examples{
\dontrun{
dyntaxa_name_from_id(206041)
dyntaxa_name_from_id(206043)
}
}
