% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_rank_from_id}
\alias{dyntaxa_rank_from_id}
\title{Get ranks for a specific taxon identifier}
\usage{
dyntaxa_rank_from_id(taxon_id)
}
\arguments{
\item{taxon_id}{taxon identifier/key}
}
\value{
rank as character or NA
}
\description{
Get ranks for a specific taxon identifier
}
\examples{
\dontrun{
dyntaxa_rank_from_id("5000001")
}
}
