% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_redlist}
\alias{dyntaxa_redlist}
\title{Get redlist data for Dyntaxa}
\usage{
dyntaxa_redlist()
}
\value{
tibble with redlist data
}
\description{
Get redlist data for Dyntaxa
}
\examples{
\dontrun{
dyntaxa_redlist()
}
}
