% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_resolve}
\alias{dyntaxa_resolve}
\title{Resolve scientific names that may be misspelled through fuzzy matching}
\usage{
dyntaxa_resolve(taxon_name)
}
\arguments{
\item{taxon_name}{taxonomic name or vector of taxonomic names}
}
\value{
tibble with the taxonomic name used in the query, a flag to indicate if it is an exact match (as opposed to a fuzzy match) and additional taxonomic data from the best taxon name that was matched (or NA if no match could be made)
}
\description{
Resolve scientific names that may be misspelled through fuzzy matching
}
\examples{
\dontrun{
dyntaxa_resolve(c("Vulpe vulp", "Canis lupu", "Vulpes vulpez"))
}
}
