% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_ng.R
\name{dyntaxa_search}
\alias{dyntaxa_search}
\title{Search dyntaxa using full text search syntax}
\usage{
dyntaxa_search(term)
}
\arguments{
\item{term}{search phrase}
}
\value{
tibble with search results
}
\description{
Use search terms except for barewords "AND", "OR" or "NOT" case sensitive.
Strings including other characters needs to be quoted. See the examples.
}
\details{
For a descripton of the search syntax, see <https://www.sqlite.org/fts5.html>
}
\examples{
\dontrun{
if(interactive()){
 dyntaxa_search("blåklocka")
 dyntaxa_search('"^italiensk*"')
 dyntaxa_search('NEAR("geting" "norsk")')$vern
 }
}
}
