% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_search_all}
\alias{dyntaxa_search_all}
\title{Search full text search index of Dyntaxa data across all data 
(identifiers, names, literature references, vernacular names etc)
using a search pattern.}
\usage{
dyntaxa_search_all(term)
}
\arguments{
\item{term}{search terms to use}
}
\value{
tibble with search hits
}
\description{
Search full text search index of Dyntaxa data across all data 
(identifiers, names, literature references, vernacular names etc)
using a search pattern.
}
\examples{
\dontrun{
dyntaxa_search_all("vitsippa AND Aronsson")
dyntaxa_search_all("Fungal Diversity")
}
}
\seealso{
\url{http://www.sqlitetutorial.net/sqlite-full-text-search/} 
for search pattern syntax details
}
