% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_search_id}
\alias{dyntaxa_search_id}
\title{Search full text search index of Dyntaxa data for taxon key(s)/identifier(s) using a search pattern}
\usage{
dyntaxa_search_id(taxon_id)
}
\arguments{
\item{taxon_id}{search term(s) specifying key/identifier}
}
\value{
tibble with columns for identfier, name, rank and status
}
\description{
Search full text search index of Dyntaxa data for taxon key(s)/identifier(s) using a search pattern
}
\examples{
\dontrun{
dyntaxa_search_id("5000001")
dyntaxa_search_id("206046 OR 206026")
dyntaxa_search_id("20604*")
}
}
\seealso{
\url{http://www.sqlitetutorial.net/sqlite-full-text-search/} for search pattern syntax details
}
