% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_search_name}
\alias{dyntaxa_search_name}
\title{Search full text search index of Dyntaxa data for scientific name using a search pattern}
\usage{
dyntaxa_search_name(taxon_name)
}
\arguments{
\item{taxon_name}{search pattern(s) for scientific name}
}
\value{
tibble with columns for identfier, name, rank and status
}
\description{
Search full text search index of Dyntaxa data for scientific name using a search pattern
}
\examples{
\dontrun{
dyntaxa_search_name("\\"Alces alces\\" OR alces")
}
}
\seealso{
\url{http://www.sqlitetutorial.net/sqlite-full-text-search/} for search pattern syntax details
}
