% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_search_vernacular}
\alias{dyntaxa_search_vernacular}
\title{Search full text search index of Dyntaxa data for vernacular name(s) using a search pattern}
\usage{
dyntaxa_search_vernacular(vernacularname)
}
\arguments{
\item{vernacularname}{search pattern(s) for vernacular name}
}
\value{
tibble with columns for identfier, name, rank and status
}
\description{
Search full text search index of Dyntaxa data for vernacular name(s) using a search pattern
}
\examples{
\dontrun{
dyntaxa_search_vernacular("lo OR fox")
dyntaxa_search_vernacular("varg*")
dyntaxa_search_vernacular("^bear*")
dyntaxa_search_vernacular("^vildsvinss*")
}
}
\seealso{
\url{http://www.sqlitetutorial.net/sqlite-full-text-search/} for search pattern syntax details
}
