% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa.R
\name{dyntaxa_synonyms}
\alias{dyntaxa_synonyms}
\title{Get synonyms given a specific taxon identifier}
\usage{
dyntaxa_synonyms(taxon_id)
}
\arguments{
\item{taxon_id}{taxon identifier/key}
}
\value{
tibble with columns for identfier, name, rank and status
}
\description{
Get synonyms given a specific taxon identifier
}
\examples{
\dontrun{
dyntaxa_synonyms(5509)
}
}
